refactor: move import and test for obfuscate_rpc_url
Move urllib.parse import to module level and relocate
test_obfuscate_rpc_url to dedicated test_blockchain_utils.py file.
